Clean analogous trade naming
